backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pandas>=2.0.0
networkx>=3.0
neo4j>=5.0
groq>=0.4.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
numpy>=1.24.0
scipy>=1.10.0
xgboost>=2.0.0
scikit-learn>=1.3.0
